+++ Code/ComfortIsland.UnitTests/Reports/BalanceReportTest.cs
@@ -130,7 +130,7 @@ namespace ComfortIsland.UnitTests.Reports
 			(
 				new[] { unit },
 				new[] { product },
-				new Dictionary<long, double>(),
+				new Dictionary<long, decimal>(),
 				new[]
 				{
 					new Document(DocumentType.Income)
